moved ChangelogService to /src (and restructured)
