--- Services/Changelog/Exception/EventHandlerNotFoundException.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace ILIAS\Changelog\Membership\Exception;
-
-
-use ilException;
-
-/**
- * Class EventHandlerNotFoundException
- * @package ILIAS\Changelog\Membership\Exception
- *
- * @author Theodor Truffer <tt@studer-raimann.ch>
- */
-class EventHandlerNotFoundException extends ilException {
-
-}
