--- Services/Changelog/Exception/RepositoryNotFoundException.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace ILIAS\Changelog\Membership\Exception;
-
-
-use ilException;
-
-/**
- * Class RepositoryNotFoundException
- * @package ILIAS\Changelog\Membership\Exception
- *
- * @author Theodor Truffer <tt@studer-raimann.ch>
- */
-class RepositoryNotFoundException extends ilException {
-
-}
