--- Services/Changelog/Exception/UnknownEventTypeException.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace ILIAS\Changelog\Membership\Exception;
-
-
-use ilException;
-
-/**
- * Class UnknownEventTypeException
- * @package ILIAS\Changelog\Membership\Exception
- *
- * @author Theodor Truffer <tt@studer-raimann.ch>
- */
-class UnknownEventTypeException extends ilException {
-
-}
