--- Services/Changelog/classes/Membership/MembershipEvent.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace ILIAS\Changelog\Membership;
-
-
-use ILIAS\Changelog\Event;
-
-/**
- * Class ChangelogMembershipEvent
- * @package ILIAS\Changelog\Membership
- *
- * @author Theodor Truffer <tt@studer-raimann.ch>
- */
-abstract class MembershipEvent implements Event {
-
-}
