--- Services/Changelog/interfaces/Event.php
@@ -1,13 +1,0 @@
-<?php
-
-
-namespace ILIAS\Changelog;
-
-
-interface Event {
-
-	/**
-	 * @return int
-	 */
-	public function getTypeId(): int;
-}
