--- Services/Changelog/interfaces/EventHandler.php
@@ -1,10 +1,0 @@
-<?php
-
-
-namespace ILIAS\Changelog;
-
-
-interface EventHandler {
-
-	public function handle(Event $changelogEvent): void;
-}
