--- Services/Changelog/interfaces/Repository.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace ILIAS\Changelog;
-
-
-/**
- * Class Repository
- * @package ILIAS\Changelog
- *
- * @author Theodor Truffer <tt@studer-raimann.ch>
- */
-interface Repository {
-
-}
