--- Services/Changelog/sql/db_update_tmp.php
@@ -1,5 +1,0 @@
-<#>
-<?php
-\ILIAS\Changelog\Infrastructure\AR\EventAR::updateDB();
-\ILIAS\Changelog\Membership\AR\MembershipEventAR::updateDB();
-?>
