+++ src/Changelog/Events/Membership/Handlers/SubscribedToCourseHandler.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace ILIAS\Changelog\Events\Membership\Handlers;
+
+
+
+use ILIAS\Changelog\Events\Membership\SubscribedToCourse;
+use ILIAS\Changelog\Interfaces\Event;
+
+/**
+ * Class SubscribedToCourseHandler
+ * @package ILIAS\Changelog\Membership\Events\Handlers
+ *
+ * @author Theodor Truffer <tt@studer-raimann.ch>
+ */
+class SubscribedToCourseHandler extends MembershipEventHandler {
+	/**
+	 * @param Event $changelogEvent
+	 */
+	public function handle(Event $changelogEvent) {
+		/** @var $changelogEvent SubscribedToCourse */
+		$this->repository->saveSubscribedToCourse($changelogEvent);
+	}
+}
