+++ src/Changelog/Events/Membership/MembershipEvent.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace ILIAS\Changelog\Events\Membership;
+
+
+use ILIAS\Changelog\Interfaces\Event;
+
+/**
+ * Class ChangelogMembershipEvent
+ * @package ILIAS\Changelog\Membership
+ *
+ * @author Theodor Truffer <tt@studer-raimann.ch>
+ */
+abstract class MembershipEvent implements Event {
+
+}
