+++ src/Changelog/Exception/EventHandlerNotFoundException.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace ILIAS\Changelog\Exception;
+
+
+use ilException;
+
+/**
+ * Class EventHandlerNotFoundException
+ * @package ILIAS\Changelog\Membership\Exception
+ *
+ * @author Theodor Truffer <tt@studer-raimann.ch>
+ */
+class EventHandlerNotFoundException extends ilException {
+
+}
