+++ src/Changelog/Exception/RepositoryNotFoundException.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace ILIAS\Changelog\Exception;
+
+
+use ilException;
+
+/**
+ * Class RepositoryNotFoundException
+ * @package ILIAS\Changelog\Membership\Exception
+ *
+ * @author Theodor Truffer <tt@studer-raimann.ch>
+ */
+class RepositoryNotFoundException extends ilException {
+
+}
