+++ src/Changelog/Exception/UnknownEventTypeException.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace ILIAS\Changelog\Exception;
+
+
+use ilException;
+
+/**
+ * Class UnknownEventTypeException
+ * @package ILIAS\Changelog\Membership\Exception
+ *
+ * @author Theodor Truffer <tt@studer-raimann.ch>
+ */
+class UnknownEventTypeException extends ilException {
+
+}
