+++ src/Changelog/Interfaces/Event.php
@@ -1,0 +1,13 @@
+<?php
+
+
+namespace ILIAS\Changelog\Interfaces;
+
+
+interface Event {
+
+	/**
+	 * @return int
+	 */
+	public function getTypeId(): int;
+}
