+++ src/Changelog/Interfaces/EventHandler.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace ILIAS\Changelog\Interfaces;
+
+
+interface EventHandler {
+
+	public function handle(Event $changelogEvent);
+}
