+++ src/Changelog/Interfaces/Repository.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace ILIAS\Changelog\Interfaces;
+
+
+/**
+ * Class Repository
+ * @package ILIAS\Changelog
+ *
+ * @author Theodor Truffer <tt@studer-raimann.ch>
+ */
+interface Repository {
+
+}
