+++ src/Changelog/sql/db_update_tmp.php
@@ -1,0 +1,5 @@
+<#>
+<?php
+\ILIAS\Changelog\Infrastructure\AR\EventAR::updateDB();
+\ILIAS\Changelog\Events\Membership\AR\MembershipEventAR::updateDB();
+?>
